+++ test/test_pt_navigator_dialog.py
@@ -6,16 +6,8 @@ from qaequilibrae.modules.public_transport_procedures.transit_navigator_dialog i
 @pytest.fixture
 def dialog(pt_project):
     dialog = TransitNavigatorDialog(pt_project)
-    dialog._testing = True
+    # dialog._testing = True
     return dialog
-
-
-def test_show_label_stops(dialog):
-    layer_labels(dialog.stops_layer, dialog.chb_label_stops)
-
-
-def test_show_label_zones(dialog):
-    layer_labels(dialog.zones_layer, dialog.chb_label_zones)
 
 
 def layer_labels(layer, chbox):
@@ -28,6 +20,14 @@ def layer_labels(layer, chbox):
     # Removes labels
     chbox.setChecked(False)
     assert not layer.labelsEnabled()
+
+
+def test_show_label_stops(dialog):
+    layer_labels(dialog.stops_layer, dialog.chb_label_stops)
+
+
+def test_show_label_zones(dialog):
+    layer_labels(dialog.zones_layer, dialog.chb_label_zones)
 
 
 def test_allow_filters(dialog):
